+++ include/types.h
@@ -114,8 +114,8 @@ typedef struct {
 typedef struct {
         c_int n;             ///< number of variables n,
         c_int m;             ///< number of constraints m
-        csc *P;              ///< P:  in csc format
-        csc *A;              ///< A: in csc format
+        csc *P;              ///< P: in csc format (size n x n)
+        csc *A;              ///< A: in csc format (size m x n)
         c_float *q;          ///< dense array for linear part of cost function (size n)
         c_float *l;          ///< dense array for lower bound (size m)
         c_float *u;          ///< dense array for upper bound (size m)
@@ -130,9 +130,9 @@ typedef struct {
          * @name These *cannot* change for multiple runs with the same call to osqp_setup
          * @{
          */
-        c_float rho;     ///< ADMM step rh
-        c_float sigma;   ///< ADMM step sigm
-        c_int scaling;   ///< boolean, heuristic data rescalin
+        c_float rho;     ///< ADMM step rho
+        c_float sigma;   ///< ADMM step sigma
+        c_int scaling;   ///< boolean, heuristic data rescaling
 
         #if EMBEDDED != 1
         c_int scaling_norm; ///< scaling norm
@@ -242,6 +242,10 @@ typedef struct {
 
         /// flag indicating whether the solve function has been run before
         c_int first_run;
+        #endif
+
+        #ifdef PRINTING
+        c_int summary_printed;  ///< Has last summary been printed? (true/false)
         #endif
 } OSQPWorkspace;