+++ src/lin_alg.c
@@ -4,31 +4,35 @@
 
 /* VECTOR FUNCTIONS ----------------------------------------------------------*/
 
-c_float vec_norm2_sq_diff(const c_float *a, const c_float *b, c_int l) {
-    c_float nmDiff = 0.0, tmp;
-    c_int i;
-    for (i = 0; i < l; ++i) {
-        tmp = (a[i] - b[i]);
-        nmDiff += tmp * tmp;
-    }
-    return nmDiff;
-}
 
 
 void vec_add_scaled(c_float *a, const c_float *b, c_int n, c_float sc) {
     c_int i;
-    for (i = 0; i < n; ++i) {
+    for (i = 0; i < n; i++) {
         a[i] += sc * b[i];
     }
 }
 
-c_float vec_norm2_sq(const c_float *v, c_int l) {
+
+c_float vec_norm_inf(const c_float *v, c_int l) {
     c_int i;
-    c_float nmsq = 0.0;
-    for (i = 0; i < l; ++i) {
-        nmsq += v[i] * v[i];
+    c_float abs_v_i;
+    c_float max = 0.0;
+    for (i = 0; i < l; i++) {
+        abs_v_i = c_absval(v[i]);
+        if ( abs_v_i > max ) max = abs_v_i;
     }
-    return nmsq;
+    return max;
+}
+
+c_float vec_norm_inf_diff(const c_float *a, const c_float *b, c_int l){
+    c_float nmDiff = 0.0, tmp;
+    c_int i;
+    for (i = 0; i < l; i++) {
+        tmp = c_absval(a[i] - b[i]);
+        if ( tmp > nmDiff ) nmDiff = tmp;
+    }
+    return nmDiff;
 }