+++ src/util.c
@@ -103,7 +103,10 @@ void print_setup_header(const OSQPData *data, const OSQPSettings *settings) {
 }
 
 
-void print_summary(OSQPInfo * info){
+void print_summary(OSQPWorkspace * work){
+    OSQPInfo * info;
+    info = work->info;
+
     c_print("%*i ", (int)strlen(HEADER[0]), (int)info->iter);
     c_print("%*.4e ", (int)HSPACE, info->obj_val);
     c_print("%*.4e ", (int)HSPACE, info->pri_res);
@@ -112,10 +115,15 @@ void print_summary(OSQPInfo * info){
     c_print("%*.2fs", 9, info->setup_time + info->solve_time);
     #endif
     c_print("\n");
+
+    work->summary_printed = 1; // Summary has been printed
 }
 
 
-void print_polish(OSQPInfo * info) {
+void print_polish(OSQPWorkspace * work) {
+    OSQPInfo * info;
+    info = work->info;
+    
     c_print("%*s ", (int)strlen(HEADER[0]), "PLSH");
     c_print("%*.4e ", (int)HSPACE, info->obj_val);
     c_print("%*.4e ", (int)HSPACE, info->pri_res);
@@ -417,11 +425,11 @@ void print_dns_matrix(c_float * M, c_int m, c_int n, const char *name)
                 for(j=0; j<n; j++) { // Cycle over columns
                         if (j < n - 1)
                                 // c_print("% 14.12e,  ", M[j*m+i]);
-                                c_print("% .5f,  ", M[j*m+i]);
+                                c_print("% .8f,  ", M[j*m+i]);
 
                         else
                                 // c_print("% 14.12e;  ", M[j*m+i]);
-                                c_print("% .5f;  ", M[j*m+i]);
+                                c_print("% .8f;  ", M[j*m+i]);
                 }
                 if (i < m - 1) {
                         c_print("\n\t");