+++ tests/solve_linsys/test_solve_linsys.h
@@ -35,8 +35,8 @@ static char * test_solveKKT(){
     solve_lin_sys(settings, p, data->test_solve_KKT_rhs);
 
     mu_assert("Linear systems solve tests: error in forming and solving KKT system!",
-              vec_norm2_sq_diff(data->test_solve_KKT_rhs, data->test_solve_KKT_x,
-                                data->test_solve_KKT_m + data->test_solve_KKT_n) < TESTS_TOL*TESTS_TOL);
+              vec_norm_inf_diff(data->test_solve_KKT_rhs, data->test_solve_KKT_x,
+                                data->test_solve_KKT_m + data->test_solve_KKT_n) < TESTS_TOL);
 
 
     // Cleanup