+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+
+Changes since the last release
+------------------------------
+* Fixed bug in index ordering of sparse matrices of Python interface
+* Changed 2-norms to inf-norms
+* Fixed code generation bug when scaling is disabled [#7](https://github.com/oxfordcontrol/osqp/issues/7)
+* Removed warnings in code-generation for standard <= C99
+* Fixed MATLAB 2015b compatibility [#6](https://github.com/oxfordcontrol/osqp/issues/6)
+
 Version 0.1.1 (11 April 2017)
 -----------------------------
 * Fixed crashes during polishing when factorization fails